backend/requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
diffusers
easyocr
pillow
python-dotenv
